backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.2
PyQtChart==5.15.7
PyQtChart-Qt5==5.15.2
sqlparse==0.5.0
tzdata==2024.1
